Rename multitest.rs file to integration_tests.rs
