+++ contracts/cw1-whitelist/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod contract;
 pub mod error;
-pub mod msg;
 #[cfg(test)]
-mod multitest;
+mod integration_tests;
+pub mod msg;
 pub mod state;
 
 pub use crate::error::ContractError;
